requirements.txt
beautifulsoup4==4.12.2
chroma-hnswlib==0.7.3
chromadb==0.4.14
langchain==0.0.319
langchainhub==0.1.13
pypdf==3.16.4
streamlit==1.27.2
tiktoken==0.5.1
python-dotenv
openai